requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.5.2
attrs==25.1.0
beautifulsoup4==4.12.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.3.1
charset-normalizer==3.4.1
click==8.1.3
colorama==0.4.6
commit-msg-hook==0.3.4
distlib==0.3.6
distro==1.9.0
Django==4.1.13
django-cors-headers==3.13.0
django-filter==22.1
django-rest-authemail==2.1.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
filelock==3.8.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.5.5
idna==3.10
importlib-metadata==4.12.0
iniconfig==1.1.1
jiter==0.8.2
joblib==1.2.0
Markdown==3.4.1
nltk==3.7
nodeenv==1.7.0
numpy==2.2.3
openai==1.63.2
outcome==1.3.0.post0
packaging==21.3
pgvector
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
psycopg2-binary==2.9.10
py==1.11.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-django==4.5.2
pytest-dotenv==0.5.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2022.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rsa==4.9
selenium==4.15.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.2
tiktoken==0.9.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.16.5
wheel==0.45.1
wsproto==1.2.0
zipp==3.8.1
dj_database_url